+++ gulpfile.js
@@ -17,6 +17,7 @@
  */
 
 import { readFileSync } from 'fs'
+import { deleteAsync } from 'del'
 
 import autoprefixer from 'gulp-autoprefixer'
 import babelify from 'babelify'
@@ -24,7 +25,7 @@ import browserify from 'browserify'
 import changed from 'gulp-changed'
 import cli from 'docker-cli-js'
 import concat from 'gulp-concat'
-import del from 'del'
+import eslint from 'gulp-eslint-new'
 import gulp from 'gulp'
 import gzip from 'gulp-gzip'
 import image from 'gulp-image'
@@ -37,7 +38,6 @@ import sassLint from 'gulp-sass-lint'
 import sort from 'gulp-sort'
 import source from 'vinyl-source-stream'
 import sourcemaps from 'gulp-sourcemaps'
-import standard from 'gulp-standard'
 import tar from 'gulp-tar'
 import zip from 'gulp-zip'
 
@@ -61,8 +61,8 @@ const dirs = {
   cache: '.cache'
 }
 
-gulp.task('clean', () => {
-  return del([
+gulp.task('clean', async () => {
+  return await deleteAsync([
     `${dirs.build}/${p.name}/**/*`,
     `${dirs.build}/${p.name}/**/.*`,
     `${dirs.build}/*zip`,
@@ -70,8 +70,8 @@ gulp.task('clean', () => {
   ])
 })
 
-gulp.task('clean-cache', () => {
-  return del([
+gulp.task('clean-cache', async () => {
+  return await deleteAsync([
     dirs.cache
   ])
 })
@@ -82,11 +82,11 @@ gulp.task('test-js', () => {
   return gulp.src(['src/js/**/*js'])
     .pipe(gulp.dest('/tmp/gulp-pre'))
     .pipe(gulp.dest('/tmp/gulp-post'))
-    .pipe(standard())
-    .pipe(standard.reporter('default', {
-      breakOnError: true,
-      quiet: true
+    .pipe(eslint({
+      overrideConfigFile: '.eslintrc.cjs'
     }))
+    .pipe(eslint.format())
+    .pipe(eslint.failAfterError())
 })
 
 gulp.task('test-sass', () => {
@@ -342,11 +342,29 @@ gulp.task('build', gulp.parallel(
 
 gulp.task('dist', gulp.parallel('build'))
 
-gulp.task('dist-test', gulp.series('clean', () => {
-  // switch to a version with no zeroes for better testing
-  p.versionEngine = p.versionEngineTest
-  return gulp.src('package.json')
-}, 'dist', () => {
+// --- testing targets ---------------------------------------------------------
+
+gulp.task('test-zips', gulp.parallel(() => {
+  return replace(gulp.src([
+    'test/data/snapshots/full/**/*'
+  ]))
+    .pipe(zip('full.zip'))
+    .pipe(gulp.dest(dirs.cache + '/snapshots'))
+}, () => {
+  return replace(gulp.src([
+    'test/data/snapshots/empty/**/*'
+  ]))
+    .pipe(zip('empty.zip'))
+    .pipe(gulp.dest(dirs.cache + '/snapshots'))
+}, () => {
+  return replace(gulp.src([
+    'test/data/snapshots/extra/**/*'
+  ]))
+    .pipe(zip('extra.zip'))
+    .pipe(gulp.dest(dirs.cache + '/snapshots'))
+}))
+
+gulp.task('dist-test', gulp.series('clean', 'test-zips', 'dist', () => {
   return replace(gulp.src([
     'test/data/server.json'
   ]))
@@ -392,15 +410,11 @@ gulp.task('release', gulp.series(
 gulp.task('release-docker', gulp.series('release', (done) => {
   const docker = new cli.Docker({ echo: true })
   docker.command('build --no-cache -t ghcr.io/ludus-leonis/freebeegee:latest .')
-    .then(() => {
-      done()
-    })
+    .then(() => { done() })
 }, (done) => {
   const docker = new cli.Docker({ echo: true })
   docker.command('tag ghcr.io/ludus-leonis/freebeegee:latest ghcr.io/ludus-leonis/freebeegee:' + p.version)
-    .then(() => {
-      done()
-    })
+    .then(() => { done() })
 }, (done) => {
   // run :api tests
   done()
@@ -450,7 +464,7 @@ gulp.task('demo-deploy-RPG', demoDeploy('RPG'))
 gulp.task('demo-deploy-Hex', demoDeploy('Hex'))
 gulp.task('demo-deploy-Tutorial', demoDeploy('Tutorial'))
 
-gulp.task('demo', gulp.series('clean', 'clean-cache', () => {
+gulp.task('demo', gulp.series('clean', () => {
   demomode = true
   site = 'https://freebeegee.org/'
   return gulp.src('tools')
@@ -472,7 +486,17 @@ gulp.task('demo', gulp.series('clean', 'clean-cache', () => {
   'demo-deploy-RPG',
   'demo-deploy-Hex',
   'demo-deploy-Tutorial'
-)))
+), async () => {
+  return await deleteAsync([
+    `${dirs.build}/demo`
+  ])
+}, () => {
+  return gulp.src([
+    `${dirs.site}/**/*`,
+    `${dirs.site}/.htaccess*`
+  ])
+    .pipe(gulp.dest(`${dirs.build}/demo`))
+}))
 
 // --- default target ----------------------------------------------------------