+++ src/js/view/room/index.mjs
@@ -518,8 +518,15 @@ function setupRoom () {
   }
   if (undefinedCount >= 4) {
     // default if store was empty
-    toggleLayer(LAYER_OTHER)
-    toggleLayer(LAYER_TOKEN)
+    if (getTemplate().layersEnabled) {
+      if (getTemplate().layersEnabled.includes(LAYER_OTHER)) toggleLayer(LAYER_OTHER)
+      if (getTemplate().layersEnabled.includes(LAYER_TOKEN)) toggleLayer(LAYER_TOKEN)
+      if (getTemplate().layersEnabled.includes(LAYER_OVERLAY)) toggleLayer(LAYER_OVERLAY)
+      if (getTemplate().layersEnabled.includes(LAYER_TILE)) toggleLayer(LAYER_TILE)
+    } else {
+      toggleLayer(LAYER_OTHER)
+      toggleLayer(LAYER_TOKEN)
+    }
   }
 
   if (getRoomPreference(PREFS.LOS)) toggleLos()
@@ -548,7 +555,7 @@ function setupRoom () {
   enableDragAndDrop('#tabletop')
 
   // load + setup content
-  setTableNo(getRoomPreference(PREFS.TABLE), false)
+  setTableNo(getRoomPreference(PREFS.TABLE) ?? getTemplate()?.table ?? 1, false)
   runStatuslineLoop()
 
   // start autosyncing after a short delay to reduce server load a bit