+++ src/js/view/room/modal/library.mjs
@@ -128,19 +128,23 @@ export function modalLibrary (xy) {
             </div>
             <div class="col-12 col-lg-6">
               <label for="upload-name">Name</label>
-              <input id="upload-name" name="name" type="text" placeholder="custom" minlength="1" maxlength="64" pattern="^[a-zA-Z0-9-]+( [a-zA-Z0-9-]+)*(, [a-zA-Z0-9-]+)?( [a-zA-Z0-9-]+)*$">
+              <input id="upload-name" name="name" type="text" placeholder="e.g. 'Sorcerer'" minlength="1" maxlength="64" pattern="^[a-zA-Z0-9-]+( [a-zA-Z0-9-]+)*(, [a-zA-Z0-9-]+)?( [a-zA-Z0-9-]+)*$">
             </div>
-            <div class="col-12 col-lg-2">
+            <div class="col-6 col-lg-1">
+              <label for="upload-w">Width</label>
+              <select id="upload-w" name="width"></select>
+            </div>
+            <div class="col-6 col-lg-1">
+              <label for="upload-h">Height</label>
+              <select id="upload-h" name="height"></select>
+            </div>
+            <div class="col-6 col-lg-2">
               <label for="upload-type">Type</label>
               <select id="upload-type" name="type"></select>
             </div>
             <div class="col-6 col-lg-2">
-              <label for="upload-w">Width</label>
-              <select id="upload-w" name="width"></select>
-            </div>
-            <div class="col-6 col-lg-2">
-              <label for="upload-h">Height</label>
-              <select id="upload-h" name="height"></select>
+              <label for="upload-h">Material</label>
+              <select id="upload-material" name="material"></select>
             </div>
             <div class="col-12">
               <label class="upload-group" for="upload-file">
@@ -267,7 +271,7 @@ function setupFooter () {
  */
 function averageColor (dataUrl) {
   _('#upload-color').value = '#808080' // color detection is async, so use interim-default
-  var image = new Image() // eslint-disable-line no-undef
+  const image = new Image() // eslint-disable-line no-undef
   image.onload = function () {
     // shrink in 2 steps for more accurate average
     const canvas8 = resizeImage(image, 8)
@@ -327,12 +331,17 @@ function modalUpload () {
     const data = {
       name: unprettyName(name.value),
       format: file.type === 'image/png' ? 'png' : 'jpg',
-      type: type,
+      type,
       w: Number(_('#upload-w').value),
       h: Number(_('#upload-h').value),
       base64: _('.upload-preview .piece').node().style.backgroundImage
-        .replace(/^[^,]*,/, '')
+        .replace(/^.*,/, '')
         .replace(/".*/, '')
+    }
+
+    const material = _('#upload-material').value
+    if (material && material !== 'none') {
+      data.tx = material
     }
 
     // set bg color
@@ -432,46 +441,106 @@ function setupTabUpload () {
   }
   height.on('change', change => updatePreview())
 
+  // height
+  const material = _('#upload-material')
+  const none = _('option').create('none')
+  none.selected = true
+  none.value = 'none'
+  material.add(none)
+  const paper = _('option').create('Paper')
+  paper.value = 'paper'
+  material.add(paper)
+  const wood = _('option').create('Wood')
+  wood.value = 'wood'
+  material.add(wood)
+  material.on('change', change => updatePreview())
+
   _('#upload-color').value = '#808080'
 
   _('#upload-file').on('change', change => updatePreview(true))
+  blob = null
 
   updatePreview()
 }
+
+let blob = null
 
 /**
  * Update the upload WYSIWYG preview based on the selected infos.
  */
 function updatePreview (parseImage = false) {
+  const file = _('#upload-file').files[0]
+
+  if (file && parseImage) {
+    const parts = splitAssetFilename(file.name)
+    if (parts.w) _('#upload-w').value = parts.w
+    if (parts.h) _('#upload-h').value = parts.h
+    if (parts.name) {
+      _('#upload-name').value = prettyName(parts.name)
+    }
+    if (['wood', 'paper'].includes(parts.tx)) {
+      _('#upload-material').value = parts.tx
+    } else {
+      _('#upload-material').value = 'none'
+    }
+    if ([LAYER_TILE, LAYER_TOKEN].includes(parts.type)) {
+      _('#upload-type').value = parts.type
+    }
+
+    const reader = new FileReader()
+    reader.addEventListener('load', event => {
+      blob = event.target.result
+
+      _('#upload-color').value = averageColor(blob)
+
+      // guess type/dimensions if no info was in filename
+      if (!parts.w) {
+        const image = new Image() // eslint-disable-line no-undef
+        image.src = blob
+        image.onload = () => {
+          const tilesize = 64 // in px
+          const aspect = image.width / image.height
+          const square = aspect > 0.9 && aspect < 1.1
+          const tilesX = Math.round(image.width / tilesize) || 1
+          const tilesY = Math.round(image.height / tilesize) || 1
+          if (square) {
+            if (tilesX <= 4) {
+              _('#upload-type').value = LAYER_TOKEN
+              _('#upload-w').value = tilesX
+              _('#upload-h').value = tilesX
+            } else {
+              _('#upload-type').value = LAYER_TILE
+              _('#upload-w').value = Math.min(tilesX, 32)
+              _('#upload-h').value = Math.min(tilesY, 32)
+            }
+          } else {
+            _('#upload-type').value = LAYER_TILE
+            _('#upload-w').value = Math.min(tilesX, 32)
+            _('#upload-h').value = Math.min(tilesY, 32)
+          }
+          updatePreviewDOM(blob)
+        }
+      } else {
+        updatePreviewDOM(blob)
+      }
+    }, false)
+    reader.readAsDataURL(file)
+  } else {
+    updatePreviewDOM(blob)
+  }
+}
+
+/**
+ * Create a new preview piece.
+ */
+function updatePreviewDOM (blob) {
   const preview = _('.modal-library .upload-preview').remove('.is-*')
   preview.innerHTML = ''
 
-  const file = _('#upload-file').files[0]
-  if (parseImage) {
-    const parts = splitAssetFilename(file.name)
-    if (_('#upload-name').value.length <= 0) { // guess defaults for form
-      if (parts.w) _('#upload-w').value = parts.w
-      if (parts.h) _('#upload-h').value = parts.h
-      if (parts.name) {
-        _('#upload-name').value = prettyName(parts.name)
-      }
-    }
-  }
-
   const type = _('#upload-type').value
+  const material = _('#upload-material').value
   const w = _('#upload-w').value
   const h = _('#upload-h').value
-
-  if (w > 16 || h > 16) {
-    preview.add('.is-deflate-4x')
-  } else if (w > 12 || h > 12) {
-    preview.add('.is-deflate-3x')
-  } else if (w > 8 || h > 8) {
-    preview.add('.is-deflate-2x')
-  } else if (w > 2 || h > 2) {
-  } else {
-    preview.add('.is-inflate-2x')
-  }
 
   // add piece to DOM
   const piece = _(`.piece.piece-${type}.is-w-${w}.is-h-${h}`).create()
@@ -480,26 +549,37 @@ function updatePreview (parseImage = false) {
   } else {
     piece.css({ '--fbg-color': '#202020' })
   }
-  preview.add(piece)
+  if (type === LAYER_TOKEN) {
+    piece.add('.has-highlight')
+  }
+  piece.css({ '--fbg-material': url(`img/material-${material}.png`) })
 
-  // set preview background image
-  if (file) {
-    const reader = new FileReader()
-    reader.addEventListener('load', event => {
-      _('.upload-preview .piece').css({
-        backgroundImage: `url("${event.target.result}")`,
-        backgroundSize: 'cover'
-      })
-      _('#upload-color').value = averageColor(event.target.result)
-    }, false)
-    reader.readAsDataURL(file)
+  if (w > 16 || h > 16) {
+    preview.add('.is-deflate-4x')
+  } else if (w > 12 || h > 12) {
+    preview.add('.is-deflate-3x')
+  } else if (w > 8 || h > 8) {
+    preview.add('.is-deflate-2x')
+  } else if (w > 2 || h > 2) {
+    // nothing
   } else {
+    preview.add('.is-inflate-2x')
+  }
+
+  if (blob) { // image loaded
+    piece.css({
+      backgroundImage: `var(--fbg-material), url("${blob}")`,
+      backgroundSize: '256px, cover'
+    })
+  } else { // show upload placeholder
     piece.css({
       backgroundImage: url('img/upload.svg'),
       backgroundSize: '25%',
       backgroundRepeat: 'no-repeat'
     })
   }
+
+  preview.add(piece)
 }
 
 /**